+++ src/main/java/com/backfcdev/managementsystem/service/ICategoryService.java
@@ -5,5 +5,5 @@ import com.backfcdev.managementsystem.dto.response.CategoryResponse;
 import com.backfcdev.managementsystem.model.Category;
 
 
-public interface ICategoryService extends ICRUD<Category, CategoryRequest, CategoryResponse, Integer> {
+public interface ICategoryService extends ICRUD<Category, CategoryRequest, CategoryResponse, Long> {
 }
