+++ src/main/java/com/backfcdev/managementsystem/service/IOrderService.java
@@ -4,5 +4,5 @@ import com.backfcdev.managementsystem.dto.request.OrderRequest;
 import com.backfcdev.managementsystem.dto.response.OrderResponse;
 import com.backfcdev.managementsystem.model.Order;
 
-public interface IOrderService extends ICRUD<Order, OrderRequest, OrderResponse, Integer>{
+public interface IOrderService extends ICRUD<Order, OrderRequest, OrderResponse, Long>{
 }
