+++ src/main/java/com/backfcdev/managementsystem/service/IProductService.java
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 import java.util.Optional;
 
-public interface IProductService extends ICRUD<Product, ProductRequest, ProductResponse, Integer> {
+public interface IProductService extends ICRUD<Product, ProductRequest, ProductResponse, Long> {
     Page<ProductResponse> findByArgs(final Optional<String> name, final Optional<Double> price, Pageable pageable);
     List<ProductResponse> findByStockLessThan(int amount);
     List<ProductResponse> getBestSellingProducts(int amount);
